Moved code from router to dataservice
